FullConfigurations

+++ AccProductsEnt.Kuriev/Entities/Accounting.cs
@@ -6,7 +6,7 @@
         public bool IsDeleted { get; set; }
 
         public  int WaybillId { get; set; }
-        public int PriceProduct { get; set; }
+        public decimal PriceProduct { get; set; }
 
 
         public ICollection<Staff> Staffs { get; set;}
